# mobile/swagger.json
{
    "schemes": [],
    "swagger": "2.0",
    "info": {
        "description": "",
        "title": "",
        "contact": {},
        "version": ""
    },
    "host": "",
    "basePath": "",
    "paths": {
...
        "/members": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "获取会员列表，支持按名称、手机号和等级筛选，并支持分页",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "会员管理"
                ],
                "summary": "获取会员列表",
                "parameters": [
                    {
                        "type": "string",
                        "description": "会员名称（模糊查询）",
                        "name": "name",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "手机号码（模糊查询）",
                        "name": "phone",
                        "in": "query"
                    },
                    {
                        "enum": [
                            "regular",
                            "silver",
                            "gold",
                            "platinum",
                            "diamond"
                        ],
                        "type": "string",
                        "description": "会员等级",
                        "name": "level",
                        "in": "query"
                    },
                    {
                        "minimum": 1,
                        "type": "integer",
                        "description": "页码，默认1",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "maximum": 100,
                        "minimum": 1,
                        "type": "integer",
                        "description": "每页数量，默认10",
                        "name": "pageSize",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功获取会员列表",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/models.PaginatedResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "items": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/models.Member"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "未授权",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "创建新的会员，需提供必要的会员信息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "会员管理"
                ],
                "summary": "创建会员",
                "parameters": [
                    {
                        "description": "会员信息",
                        "name": "member",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Member"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "创建成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/models.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/models.Member"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "请求参数错误或手机号已被注册",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "未授权",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/members/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "根据ID获取会员的详细信息，包括基本信息、等级、积分等",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "会员管理"
                ],
                "summary": "获取会员详情",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "会员ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功获取会员详情",
                        "schema": {
                            "$ref": "#/definitions/models.Member"
                        }
                    },
                    "401": {
                        "description": "未授权",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "会员不存在",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "更新现有会员的信息，需提供完整的会员信息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "会员管理"
                ],
                "summary": "更新会员",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "会员ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "会员信息",
                        "name": "member",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Member"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "更新成功",
                        "schema": {
                            "$ref": "#/definitions/models.Member"
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "未授权",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "会员不存在",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "删除指定ID的会员，此操作不可逆",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "会员管理"
                ],
                "summary": "删除会员",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "会员ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "删除成功",
                        "schema": {
                            "$ref": "#/definitions/models.APIResponse"
                        }
                    },
                    "401": {
                        "description": "未授权",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "会员不存在",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "models.APIResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "状态码",
                    "type": "integer",
                    "example": 200
                },
                "data": {
                    "description": "响应数据"
                },
                "error": {
                    "description": "错误信息，仅在发生错误时返回",
                    "type": "string",
                    "example": "错误信息"
                },
                "message": {
                    "description": "响应消息",
                    "type": "string",
                    "example": "操作成功"
                }
            }
        },
        "models.ConsumptionLevel": {
            "type": "string",
            "enum": [
                "low",
                "medium",
                "high",
                "luxury"
            ],
            "x-enum-comments": {
                "High": "高",
                "Low": "低",
                "Luxury": "奢侈",
                "Medium": "中"
            },
            "x-enum-varnames": [
                "Low",
                "Medium",
                "High",
                "Luxury"
            ]
        },
...
        "models.ErrorResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "错误码",
                    "type": "integer",
                    "example": 400
                },
                "error": {
                    "description": "详细错误信息，仅在开发环境返回",
                    "type": "string"
                },
                "message": {
                    "description": "错误消息",
                    "type": "string",
                    "example": "参数错误"
                }
            }
        },
...
        "models.Member": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string"
                },
                "birthday": {
                    "type": "string"
                },
                "bodyHeight": {
                    "description": "体型数据",
                    "type": "integer"
                },
                "bodyWeight": {
                    "description": "体重(kg)",
                    "type": "integer"
                },
                "bustSize": {
                    "description": "胸围(cm)",
                    "type": "integer"
                },
                "consumptionLevel": {
                    "description": "消费能力",
                    "allOf": [
                        {
                            "$ref": "#/definitions/models.ConsumptionLevel"
                        }
                    ]
                },
                "createdAt": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "favoriteCategories": {
                    "description": "喜欢的品类，逗号分隔",
                    "type": "string"
                },
                "favoriteColors": {
                    "description": "喜欢的颜色，逗号分隔",
                    "type": "string"
                },
                "gender": {
                    "type": "string"
                },
                "hipSize": {
                    "description": "臀围(cm)",
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "inseam": {
                    "description": "内缝长度(cm)",
                    "type": "integer"
                },
                "lastPurchaseDate": {
                    "type": "string"
                },
                "level": {
                    "$ref": "#/definitions/models.MemberLevel"
                },
                "name": {
                    "type": "string"
                },
                "note": {
                    "description": "备注",
                    "type": "string"
                },
                "openID": {
                    "description": "微信 OpenID",
                    "type": "string"
                },
                "phone": {
                    "type": "string"
                },
                "points": {
                    "description": "积分",
                    "type": "integer"
                },
                "shoulderWidth": {
                    "description": "肩宽(cm)",
                    "type": "integer"
                },
                "stylePreference": {
                    "description": "偏好数据",
                    "allOf": [
                        {
                            "$ref": "#/definitions/models.StylePreference"
                        }
                    ]
                },
                "totalSpent": {
                    "description": "累计消费金额",
                    "type": "number"
                },
                "updatedAt": {
                    "type": "string"
                },
                "waistSize": {
                    "description": "腰围(cm)",
                    "type": "integer"
                }
            }
        },
        "models.MemberLevel": {
            "type": "string",
            "enum": [
                "regular",
                "silver",
                "gold",
                "platinum",
                "diamond"
            ],
            "x-enum-comments": {
                "Diamond": "钻石会员",
                "Gold": "黄金会员",
                "Platinum": "铂金会员",
                "Regular": "普通会员",
                "Silver": "白银会员"
            },
            "x-enum-varnames": [
                "Regular",
                "Silver",
                "Gold",
                "Platinum",
                "Diamond"
            ]
        },
        "models.PaginatedResponse": {
            "type": "object",
            "properties": {
                "items": {
                    "description": "数据项列表"
                },
                "page": {
                    "description": "当前页码",
                    "type": "integer",
                    "example": 1
                },
                "pageSize": {
                    "description": "每页记录数",
                    "type": "integer",
                    "example": 10
                },
                "total": {
                    "description": "总记录数",
                    "type": "integer",
                    "example": 100
                }
            }
        },
...
        "models.StylePreference": {
            "type": "string",
            "enum": [
                "casual",
                "formal",
                "sportswear",
                "vintage",
                "minimalist",
                "romantic",
                "bohemian",
                "street"
            ],
            "x-enum-comments": {
                "Bohemian": "波西米亚",
                "Casual": "休闲",
                "Formal": "正式",
                "Minimalist": "极简",
                "Romantic": "浪漫",
                "Sportswear": "运动",
                "Street": "街头",
                "Vintage": "复古"
            },
            "x-enum-varnames": [
                "Casual",
                "Formal",
                "Sportswear",
                "Vintage",
                "Minimalist",
                "Romantic",
                "Bohemian",
                "Street"
            ]
        },
...
    }
}